+++ bugster/commands/update.py
@@ -1,4 +1,3 @@
-import os
 import subprocess
 
 import yaml
@@ -10,23 +9,13 @@ from bugster.analyzer.core.app_analyzer.utils.get_tree_structure import (
     get_gitignore,
 )
 from bugster.libs.services.test_cases_service import TestCasesService
+from bugster.libs.utils.files import get_all_files
 
 console = Console()
 
 
 DIR_PATH = "."
 TEST_CASES_PATH = "test_cases"
-
-
-def get_all_files(directory):
-    """Get all files in a directory."""
-    file_paths = []
-
-    for root, dirs, files in os.walk(directory):
-        for file in files:
-            file_paths.append(os.path.join(root, file))
-
-    return file_paths
 
 
 def find_pages_that_use_file(file: str) -> list[str]:
